pyproject.toml
[tool.poetry]
name = "asymc_api_tests"
version = "0.1.0"
description = "Pytest + pytest-asyncio"
authors = ["Igor Balagurov <balagurovi@yandex.ru>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
pytest = "^5.3.5"
allure-pytest = "^2.8.10"
aiohttp = "^3.6.2"
pytest-trio = "^0.5.2"
pytest-subtests = "^0.3.0"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
flake8 = "^3.7.9"
flake8-builtins = "^1.4.2"
flake8-bugbear = "^20.1.4"
flake8-coding = "^1.3.2"
flake8-comprehensions = "^3.2.2"
flake8-eradicate = "^0.2.4"
flake8-logging-format = "^0.6.0"
pep8-naming = "^0.9.1"
flake8-string-format = "^0.2.3"
flake8-pep3101 = "^1.3.0"
flake8-bandit = "^2.1.2"
flake8-print = "^3.1.4"
flake8-debugger = "^3.2.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
